+++ multimodal_challenge/util.py
@@ -1,14 +1,22 @@
+from packaging import version
 from json import loads
 from typing import List, Dict
+from pkg_resources import get_distribution
 from os.path import join
 from tdw.librarian import SceneLibrarian
 from tdw.tdw_utils import TDWUtils
+from tdw.version import __version__
+from tdw.release.pypi import PyPi
 from multimodal_challenge.paths import TARGET_OBJECTS_PATH, OBJECT_INIT_DIRECTORY, SCENE_LIBRARY_PATH, \
     ASSET_BUNDLES_DIRECTORY
 from multimodal_challenge.multimodal_object_init_data import MultiModalObjectInitData
 
 # A list of the names of target objects models.
 TARGET_OBJECTS: List[str] = TARGET_OBJECTS_PATH.read_text(encoding="utf-8").split("\n")
+# The required version of TDW.
+TDW_REQUIRED_VERSION = "1.8.19.0"
+# The required version of Magnebot.
+MAGNEBOT_REQUIRED_VERSION = "1.2.1"
 
 
 def get_scene_librarian() -> SceneLibrarian:
@@ -71,3 +79,54 @@ def get_trial_filename(trial: int) -> str:
     """
 
     return TDWUtils.zero_padding(trial, 5)
+
+
+def check_pip_version() -> bool:
+    """
+    Check the version of TDW and Magenbot.
+
+    :return: True if both the tdw and magnebot pip modules are at the correct version.
+    """
+
+    ok = True
+    # Check the version of TDW.
+    # Use the __version__ variable because it's more likely to be accurate.
+    # The version returned by get_distribution can be wrong if this is a test branch.
+    # We don't really need to worry about this re: magnebot because that repo isn't updated as frequently.
+    if version.parse(TDW_REQUIRED_VERSION) != version.parse(__version__):
+        print(f"WARNING! You have tdw {__version__} but you need tdw {TDW_REQUIRED_VERSION}. "
+              f"To install the correct version:"
+              f"\nIf you installed tdw from the GitHub repo (pip3 install -e .): "
+              f"git checkout v{PyPi.strip_post_release(TDW_REQUIRED_VERSION)}"
+              f"\nIf you installed tdw from PyPi (pip3 install tdw): "
+              f"pip3 install tdw=={TDW_REQUIRED_VERSION}")
+        ok = False
+    magnebot_installed_version = get_distribution("magnebot").version
+    if version.parse(MAGNEBOT_REQUIRED_VERSION) != version.parse(magnebot_installed_version):
+        print(f"WARNING! You have magnebot {magnebot_installed_version} but you need magnebot {MAGNEBOT_REQUIRED_VERSION}. "
+              f"To install the correct version:"
+              f"\nIf you installed tdw from the GitHub repo (pip3 install -e .): "
+              f"git checkout {MAGNEBOT_REQUIRED_VERSION}"
+              f"\nIf you installed tdw from PyPi (pip3 install magnebot): "
+              f"pip3 install magnebot=={MAGNEBOT_REQUIRED_VERSION}")
+        ok = False
+    return ok
+
+
+def check_build_version(build_version: str) -> bool:
+    """
+    Check the version of the build.
+
+    :param build_version: The version of the build.
+
+    :return: True if this is the correct version of the build.
+    """
+
+    tdw_required_version_stripped = PyPi.strip_post_release(TDW_REQUIRED_VERSION)
+    if version.parse(build_version) != version.parse(tdw_required_version_stripped):
+        print(f"WARNING! You are using TDW build {build_version} but you need TDW build {tdw_required_version_stripped}. "
+              f"\nDownload and extract from here: "
+              f"https://github.com/threedworld-mit/tdw/releases/tag/v{tdw_required_version_stripped}")
+        return False
+    else:
+        return True